Major refactoring to split the core file ingot easiest understandable parts. Main difference is a new file named node that include most of the different puzzle pieces that create things in the model. Core file is more specialized in building the graph.
I also prepared some subfolders to include standard devices, sensors, etc in a readable way.

I've also reconfigured the tests so pytest can be used. The bash script doesn't play well with Windows. It's actually very simple for a lot of test and I just consider OK th efact that nothing is broken. For some test, I used the with raise option of pytest to detect the exception.

I know this is a big chunk, but this will allow more flexibility when adding new stuff.